chore: rename to CodeBlock
