--- packages/api-reference/src/components/Code/index.ts
@@ -1,1 +1,0 @@
-export { default as Code } from './Code.vue'
