+++ packages/api-reference/src/components/CodeBlock/index.ts
@@ -1,0 +1,1 @@
+export { default as CodeBlock } from './CodeBlock.vue'
